Web_Application/requirements.txt
Django==2.2
djangorestframework==3.10.3
mongoengine==0.18.2
pymongo==3.9.0
requests==2.22.0
